Move prune logic to image package
